Rewrite EventHandle for late initialization
Temporarily break event registering in preparation for separating Events and Dispatcher
