+++ omni-led-lib/src/events/mod.rs
@@ -1,3 +1,4 @@
+pub mod event_handle;
 pub mod event_loop;
 pub mod event_queue;
 pub mod events;
